Blend sorting into filtering
Since filters already collect parameters in an array-like workspace
allocation, sorting them only requires a call to qsort(3) and it's a
problem solved.

It means that a filter can do all of cleaning, filtering and sorting
at the same time with no additional workspace penalty. Again, the
`querystring.sort` function becomes a shorthand of a static filter
and we can finally get rid of the dedicated error-prone function.