Overhaul of the build system for a greater good
This is an attempt at making the VMOD building blocks reusable and
removing as much boilerplates as autotools allow us (which isn't much
but still better than before this commit).

The first action is to remove reusable parts out of `configure.ac`:

- a dedicated m4 file takes care of the PKG_CHECK_VAR compat hack
- a reusable m4 file for VMODs building is introduced

In configure.ac:

    AX_VMOD_BUILD(querystring)

In src/Makefile.am:

    vmod_LTLIBRARIES = libvmod_querystring.la

    dist_man_MANS = vmod_querystring.3

    libvmod_querystring_la_SOURCES = \
        vmod_querystring.c

    nodist_libvmod_querystring_la_SOURCES = \
        vcc_querystring_if.c \
        vcc_querystring_if.h

    @BUILD_VMOD_QUERYSTRING@

It is not possible to hide the `nodist_..._SOURCES` even though there's
a recurring pattern here. Just like $(VMOD_INCLUDES) are declared for
the whole Makefile, it is now the case for $(VMOD_LDFLAGS) too. Hence
the lack of LDFLAGS in the example above.

The reusable VMOD rules make use of Varnish 4.1.2's ability to choose
the name of the VCC interface files. It should also be reusable by
multi-module projects. Consequently, it expects the vcc file to
incorporate the VMOD's name.

While at it, let's make the `vmodtool` invocaction more friendly with
silent rules, and why not turn the whole build silent too?

Other minor changes include overall polish and consistency efforts.

Current status: crossing fingers hoping that it doesn't break too much
on supported platforms.